requirements.txt
Django==1.11.16
mysqlclient==1.3.13
django-crispy-forms==1.7.2
django-json-rpc==0.7.2
